EditMode & DAO System Change
Started implementing logic for EditMode. Adding conditionally rendered
display and edit ui:fragments. Buttons for accept->update and cancel.

Reorganized DAOs. Now all implement same Interface used for Injections.
If special case should come, special interface needs to be programmed
and used for injection of the special case DAO. This allowed to remove
"empty" Interfaces for existing DAOs. Possible future update to Service
on the same terms.

Appended lokale to properties files name labels -> labelsDE.

Removed all ManagedProperty(JSF) from Spring managed Beans.

+++ kapsel/src/main/java/de/kapsel/produkt/entities/Arbeitsschritt.java
@@ -21,7 +21,7 @@ public class Arbeitsschritt implements DTItem, Serializable, Comparable<Arbeitss
 	private long id;
 	private int position=0;
 	private String name="";
-	private double zeit=0;
+	private int zeit=0;
 	private String notiz="";
 	private ETypes.AschrittT typ;
 	private Werkzeug werkzeug;
@@ -54,10 +54,10 @@ public class Arbeitsschritt implements DTItem, Serializable, Comparable<Arbeitss
 	}
 	
 	@Column(name="zeit")
-	public double getZeit() {
+	public int getZeit() {
 		return zeit;
 	}
-	public void setZeit(double zeit) {
+	public void setZeit(int zeit) {
 		this.zeit = zeit;
 	}
